fix(paymob): loose-cast subscription->user_id before comparing to Auth::id()
Root cause of the real 403 (not the WAF false alarm from the previous
commit — that one was real too, but this is the actual blocker): all three
owner-authorization checks used a strict === between $subscription->user_id
and Auth::id(). If the DB driver's PDO config returns an integer column as
a numeric string in one environment but not another (a real, common
config difference, unrelated to anything in this app's own code), a
genuinely-owning logged-in user fails a strict === and gets a 403 for
their own subscription. Confirmed against production: incognito, freshly
logged in, own manual-transfer submission — still 403'd on the exact
result page it should have shown.

No behavior change for the correct case, no security weakening (an int
cast of a non-owner's id still can't collide with the real owner's).

+++ app/Http/Controllers/Web/PaymobCallbackController.php
@@ -68,7 +68,15 @@ class PaymobCallbackController extends Controller
      */
     private function authorize(Request $request, Subscription $subscription): void
     {
-        $authorized = (Auth::check() && $subscription->user_id === Auth::id())
+        // Cast both sides before comparing — some DB driver configurations
+        // (PDO::ATTR_EMULATE_PREPARES differences between environments) can
+        // return an integer column as a numeric STRING; Auth::id() is always
+        // an int, so a strict === between the two would silently reject the
+        // real owner. Confirmed against a live 403 for a genuinely logged-in
+        // owner. Never null-equals-null: (int) null is 0, and Auth::id() is
+        // never 0 while Auth::check() is true, so a guest subscription can't
+        // accidentally match here.
+        $authorized = (Auth::check() && (int) $subscription->user_id === (int) Auth::id())
             || (!Auth::check()
                 && $subscription->guest_token
                 && hash_equals($subscription->guest_token, (string) $request->query('guest_token', '')));
